Continue on LDAP
Removed saslauth and using mod_auth_ldap instead. No prosody-0.12,
instead using Ubuntu's prosody package (0.11 for Ubuntu 22). Some
comments about cyrus need to be removed (and the commented-out cyrus
module).

Experimented and it asks for login but it immediately responds with
"Incorrect username or password". If I attempt to use a domain in the
username (such as "user.0@example.com") it returns some other error
(unless I use the domain "meet.jitsi"). Need to check what exactly is
the LDAP query it makes.